created crud section

+++ src/main/java/com/meli/freshWarehouse/dto/SectionDto.java
@@ -2,12 +2,13 @@ package com.meli.freshWarehouse.dto;
 
 import lombok.Builder;
 import lombok.Data;
+
 @Data
 @Builder
 public class SectionDto {
 
     private String name;
-    private int availableSpace;
-    private int idWarehouse;
+    private Integer availableSpace;
+    private Long idWarehouse;
 
 }
